Cleanup and fix test
